Reworked entities synchronization

+++ src/resources/position_update.rs
@@ -1,4 +1,4 @@
 use crate::utils::Position;
 use std::collections::HashMap;
 
-pub type PositionUpdateResource = HashMap<u16, Position>;
+pub type PositionUpdateResource = HashMap<u32, Position>;
